add androidバージョン設定
