+++ lib/ui/home/view/home_page.dart
@@ -1,7 +1,5 @@
 import 'package:animeishi/ui/auth/view/auth_page.dart';
 import 'package:flutter/material.dart';
-import 'package:firebase_auth/firebase_auth.dart' hide EmailAuthProvider;
-import 'package:firebase_ui_auth/firebase_ui_auth.dart';
 
 class HomePage extends StatelessWidget {
   @override
